--- a/include/dsu.h
+++ b/include/dsu.h
@@ -1,43 +1,46 @@
 #pragma once
 #include <vector>
 
+template<class T>
+struct DSUMergeRet {
+  bool merged; // true if a merge actually occured
+  T root; // new root
+  T child; // new child of root
+};
+
 template <class T>
 class DisjointSetUnion {
   std::vector<T> parent;
   std::vector<T> size;
 public:
-  DisjointSetUnion();
-  DisjointSetUnion(T n);
-
-  void link(T i, T j);
-  T find_set(T i);
-  void union_set(T i, T j);
-};
-
-template <class T>
-DisjointSetUnion<T>::DisjointSetUnion() {}
-
-template <class T>
-DisjointSetUnion<T>::DisjointSetUnion(T n) : parent(n), size(n, 1) {
-  for (T i = 0; i < n; i++) {
-    parent[i] = i;
+  DisjointSetUnion(T n) : parent(n), size(n, 1) {
+    for (T i = 0; i < n; i++) {
+      parent[i] = i;
+    }
   }
-}
 
-template <class T>
-void DisjointSetUnion<T>::link(T i, T j) {
-  if (size[i] < size[j]) std::swap(i,j);
-  parent[j] = i;
-  size[i] += size[j];
-}
+  inline T find_root(T u) {
+    while(parent[parent[u]] != u) {
+      parent[u] = parent[parent[u]];
+      u = parent[u];
+    }
+    return u;
+  }
+  inline DSUMergeRet<T> merge(T u, T v) {
+    T a = find_root(u);
+    T b = find_root(v);
+    if (a == b) return {false, 0, 0};
 
-template <class T>
-T DisjointSetUnion<T>::find_set(T i) {
-  if (parent[i] == i) return i;
-  return parent[i] = find_set(parent[i]);
-}
+    if (size[a] < size[b]) std::swap(a,b);
+    parent[b] = a;
+    size[a] += size[b];
+    return {true, a, b};
+  }
 
-template <class T>
-void DisjointSetUnion<T>::union_set(T i, T j) {
-  link(find_set(i), find_set(j));
-}
+  inline void reset() {
+    for (T i = 0; i < parent.size(); i++) {
+      parent[i] = i;
+      size[i] = 1;
+    }
+  }
+};