--- a/include/supernode.h
+++ b/include/supernode.h
@@ -84,8 +84,7 @@ public:
   static inline void configure(uint64_t n, vec_t sketch_fail_factor=100) {
     Sketch::configure(n*n, sketch_fail_factor);
     bytes_size = sizeof(Supernode) + log2(n)/(log2(3)-1) * Sketch::sketchSizeof() - sizeof(char);
-    serialized_size = bytes_size - log2(n)/(log2(3)-1) * (sizeof(Sketch) - sizeof(char));
-    serialized_size -= sizeof(Supernode) - sizeof(char);
+    serialized_size = log2(n)/(log2(3)-1) * Sketch::serialized_size();
   }
 
   static inline size_t get_size() {