--- a/include/test/file_graph_verifier.h
+++ b/include/test/file_graph_verifier.h
@@ -12,7 +12,7 @@
 class FileGraphVerifier : public GraphVerifier {
   std::vector<std::set<node_id_t>> kruskal_ref;
   std::vector<std::set<node_id_t>> boruvka_cc;
-  std::vector<std::set<node_id_t>> det_graph;
+  std::vector<std::vector<bool>> adj_matrix;
   DisjointSetUnion<node_id_t> sets;
 
 public: