--- a/include/types.h
+++ b/include/types.h
@@ -7,11 +7,20 @@ static const auto& vec_hash = XXH32;
 static const auto& col_hash = XXH64;
 
 // Is a stream update an insertion or a deletion
-// NXT_QUERY: special type to indicate getting more data would cross a query boundary
-// END_OF_FILE: special type to indicate that there is no more data in the stream
+// BREAKPOINT: special type that indicates that a break point has been reached
+// a break point may be either the end of the stream or the index of a query
 enum UpdateType {
   INSERT = 0,
   DELETE = 1,
-  NXT_QUERY = 2,
-  END_OF_FILE = 3
-}; 
+  BREAKPOINT = 2
+};
+
+struct Edge {
+  node_id_t src = 0;
+  node_id_t dst = 0;
+};
+
+struct GraphUpdate {
+  Edge edge;
+  UpdateType type;
+};