--- a/src/graph.cpp
+++ b/src/graph.cpp
@@ -307,9 +307,7 @@ std::vector<std::set<node_id_t>> Graph::boruvka_emulation(bool make_copy) {
     std::rethrow_exception(std::current_exception());
   }
   cleanup_copy();
-#ifdef USE_EAGER_DSU
   dsu_valid = true;
-#endif // USE_EAGER_DSU
 
   auto retval = cc_from_dsu();
   cc_alg_end = std::chrono::steady_clock::now();
@@ -345,7 +343,6 @@ void Graph::restore_from_disk(const std::vector<node_id_t>& ids_to_restore) {
 }
 
 std::vector<std::set<node_id_t>> Graph::connected_components(bool cont) {
-#ifdef USE_EAGER_DSU
   // DSU check before calling force_flush()
   if (dsu_valid && cont
 #ifdef VERIFY_SAMPLES_F
@@ -367,7 +364,6 @@ std::vector<std::set<node_id_t>> Graph::connected_components(bool cont) {
 #endif
     return retval;
   }
-#endif // USE_EAGER_DSU
 
   flush_start = std::chrono::steady_clock::now();
   gts->force_flush(); // flush everything in guttering system to make final updates
@@ -426,7 +422,6 @@ std::vector<std::set<node_id_t>> Graph::cc_from_dsu() {
 }
 
 bool Graph::point_query(node_id_t a, node_id_t b) {
-#ifdef USE_EAGER_DSU
   // DSU check before calling force_flush()
   if (dsu_valid) {
     cc_alg_start = flush_start = flush_end = std::chrono::steady_clock::now();
@@ -441,7 +436,6 @@ bool Graph::point_query(node_id_t a, node_id_t b) {
     cc_alg_end = std::chrono::steady_clock::now();
     return retval;
   }
-#endif // USE_EAGER_DSU
 
 
   flush_start = std::chrono::steady_clock::now();