--- a/src/util.cpp
+++ b/src/util.cpp
@@ -7,14 +7,14 @@
 typedef uint32_t ul;
 typedef uint64_t ull;
 
-const ull ULLMAX = std::numeric_limits<ull>::max();
-const uint8_t num_bits = sizeof(node_id_t) * 8;
+constexpr ull ULLMAX = std::numeric_limits<ull>::max();
+constexpr uint8_t num_bits = sizeof(node_id_t) * 8;
 
 unsigned long long int double_to_ull(double d, double epsilon) {
   return (unsigned long long) (d + epsilon);
 }
 
-uint64_t nondirectional_non_self_edge_pairing_fn(uint32_t i, uint32_t j) {
+edge_id_t nondirectional_non_self_edge_pairing_fn(node_id_t i, node_id_t j) {
   // swap i,j if necessary
   if (i > j) {
     std::swap(i,j);
@@ -30,7 +30,7 @@ uint64_t nondirectional_non_self_edge_pairing_fn(uint32_t i, uint32_t j) {
   return i+j;
 }
 
-std::pair<uint32_t , uint32_t> inv_nondir_non_self_edge_pairing_fn(uint64_t idx) {
+Edge inv_nondir_non_self_edge_pairing_fn(uint64_t idx) {
   // we ignore possible overflow
   ull eidx = 8ull*idx + 1ull;
   eidx = sqrt(eidx)+1ull;
@@ -38,19 +38,19 @@ std::pair<uint32_t , uint32_t> inv_nondir_non_self_edge_pairing_fn(uint64_t idx)
   ull i,j = (ull) eidx;
   if ((j & 1ull) == 0ull) i = idx-(j>>1ull)*(j-1ull);
   else i = idx-j*((j-1ull)>>1ull);
-  return {i, j};
+  return {(node_id_t)i, (node_id_t)j};
 }
 
-ull concat_pairing_fn(ul i, ul j) {
+edge_id_t concat_pairing_fn(node_id_t i, node_id_t j) {
   // swap i,j if necessary
   if (i > j) {
     std::swap(i,j);
   }
-  return ((ull)i << num_bits) | j;
+  return ((edge_id_t)i << num_bits) | j;
 }
 
-std::pair<ul, ul> inv_concat_pairing_fn(ull idx) {
-  ul j = idx & 0xFFFFFFFF;
-  ul i = idx >> num_bits;
+Edge inv_concat_pairing_fn(ull idx) {
+  node_id_t j = idx & 0xFFFFFFFF;
+  node_id_t i = idx >> num_bits;
   return {i, j};
 }