--- a/tools/statistical_testing/graph_testing.cpp
+++ b/tools/statistical_testing/graph_testing.cpp
@@ -18,7 +18,7 @@ static inline int do_run() {
         g.update({{a, b}, INSERT});
       } else g.update({{a, b}, DELETE});
     }
-    g.set_verifier(std::make_unique<FileGraphVerifier>("./cumul_sample.txt"));
+    g.set_verifier(std::make_unique<FileGraphVerifier>(n, "./cumul_sample.txt"));
     try {
         g.connected_components();
     } catch (std::exception const &err) {
@@ -93,6 +93,4 @@ int main() {
         trial_list.clear();
         out.close();
     }
-    
-    
 }