DSU updates to use path halving
Serialized size getters for Supernode and Sketch