@etwest 2023-02-07
Now that we have confirmation that this new design isn't a performance detriment, I'd like to get these changes merged.